[#66] feature: keyword, amount 물건 입력 경고 모달창 추가
store에 물건이 제대로 담겨있지 않은 상황 처리
기존 개발중일때 사용하던 title 삭제
경고 모달창 추가, 확인 누르면 /list 이동
